requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
transformers>=4.30.0
numpy
scikit-learn
tqdm
qdrant-client
matplotlib
seaborn
Pillow
kaggle  # For downloading NUS-WIDE dataset
scipy